+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/key/view/MCPToolPermissionView.kt
@@ -1,7 +1,12 @@
 package ai.mcpdirect.studio.app.key.view
 
+import ai.mcpdirect.mcpdirectstudioapp.getPlatform
+import ai.mcpdirect.studio.app.Screen
 import ai.mcpdirect.studio.app.compose.StudioActionBar
+import ai.mcpdirect.studio.app.compose.TooltipIconButton
+import ai.mcpdirect.studio.app.generalViewModel
 import ai.mcpdirect.studio.app.model.AIPortServiceResponse
+import ai.mcpdirect.studio.app.model.account.AIPortUser
 import ai.mcpdirect.studio.app.model.aitool.AIPortTool
 import ai.mcpdirect.studio.app.model.aitool.AIPortToolAccessKey
 import ai.mcpdirect.studio.app.model.aitool.AIPortToolAgent
@@ -10,28 +15,45 @@ import ai.mcpdirect.studio.app.model.aitool.AIPortToolPermission
 import ai.mcpdirect.studio.app.model.repository.StudioRepository
 import ai.mcpdirect.studio.app.model.repository.ToolRepository
 import androidx.compose.foundation.background
+import androidx.compose.foundation.hoverable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.collectIsHoveredAsState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.material3.Badge
+import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.OutlinedCard
+import androidx.compose.material3.RichTooltip
+import androidx.compose.material3.RichTooltipColors
 import androidx.compose.material3.Text
+import androidx.compose.material3.TooltipAnchorPosition
+import androidx.compose.material3.TooltipBox
+import androidx.compose.material3.TooltipDefaults
+import androidx.compose.material3.rememberTooltipState
 import androidx.compose.runtime.*
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -43,8 +65,14 @@ import kotlinx.coroutines.launch
 import mcpdirectstudioapp.composeapp.generated.resources.Res
 import mcpdirectstudioapp.composeapp.generated.resources.close
 import mcpdirectstudioapp.composeapp.generated.resources.collapse_all
+import mcpdirectstudioapp.composeapp.generated.resources.description
 import mcpdirectstudioapp.composeapp.generated.resources.expand_all
+import mcpdirectstudioapp.composeapp.generated.resources.logout
+import mcpdirectstudioapp.composeapp.generated.resources.password
+import mcpdirectstudioapp.composeapp.generated.resources.reset_settings
 import mcpdirectstudioapp.composeapp.generated.resources.search
+import mcpdirectstudioapp.composeapp.generated.resources.settings
+import mcpdirectstudioapp.composeapp.generated.resources.shield_toggle
 import org.jetbrains.compose.resources.painterResource
 
 class MCPServerToolPermissionViewModel : ViewModel() {
@@ -73,18 +101,26 @@ class MCPServerToolPermissionViewModel : ViewModel() {
         }
     }
 }
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun MCPServerToolPermissionView(
-    toolMaker: AIPortToolMaker
+    toolMaker: AIPortToolMaker,
 ){
-    val viewModel = MCPServerToolPermissionViewModel()
+    val viewModel by rememberSaveable(toolMaker.id) { mutableStateOf(MCPServerToolPermissionViewModel()) }
     val localToolAgent by StudioRepository.localToolAgent.collectAsState()
     val tools by viewModel.tools.collectAsState()
     LaunchedEffect(toolMaker){
         viewModel.toolMaker(toolMaker)
     }
     OutlinedCard(Modifier.fillMaxWidth()) {
-        StudioActionBar(toolMaker.name) {
+        StudioActionBar(toolMaker.name, navigationIcon = {
+            Checkbox(checked = false,onCheckedChange = {}, Modifier.size(32.dp))
+        }) {
+            Icon(painterResource(Res.drawable.shield_toggle),contentDescription = null)
+            Spacer(Modifier.weight(1f))
+            IconButton(onClick = {}){
+                Icon(painterResource(Res.drawable.reset_settings),contentDescription = null, Modifier.size(20.dp))
+            }
             IconButton(onClick = { viewModel.expanded = !viewModel.expanded },modifier = Modifier.size(32.dp)) {
                 val icon = if(viewModel.expanded) Res.drawable.collapse_all else Res.drawable.expand_all
                 Icon(painterResource(icon),contentDescription = null, Modifier.size(20.dp))
@@ -96,12 +132,39 @@ fun MCPServerToolPermissionView(
             verticalArrangement = Arrangement.spacedBy(8.dp)
         ){
             tools.forEach { tool ->
-                OutlinedButton(onClick = {},
-                    Modifier.height(32.dp),
-                    contentPadding = PaddingValues(horizontal = 8.dp),
+                var checked by remember { mutableStateOf(false) }
+                val interactionSource = remember { MutableInteractionSource() }
+                val isHovered by interactionSource.collectIsHoveredAsState()
+                Button(
+                    modifier = Modifier.height(32.dp).hoverable(interactionSource),
+                    shape = if(checked) ButtonDefaults.shape else ButtonDefaults.textShape,
+                    colors = if(checked) ButtonDefaults.buttonColors() else ButtonDefaults.textButtonColors(),
+                    elevation = if(checked) ButtonDefaults.buttonElevation() else null,
+                    onClick = {checked=!checked},
+                    contentPadding = PaddingValues(0.dp),
                 ) {
-                    Text(tool.name)
+                    Box(contentAlignment = Alignment.CenterEnd) {
+                        Text(tool.name, Modifier.padding(horizontal = 8.dp))
+                        if (isHovered) {
+                            val colors = TooltipDefaults.richTooltipColors()
+                            IconButton(
+                                onClick = {}, Modifier.size(24.dp),
+                                colors = IconButtonDefaults.iconButtonColors().copy(
+                                    containerColor = colors.containerColor,
+                                    contentColor = colors.contentColor,
+                                ),
+                            ) {
+                                Icon(
+                                    painterResource(Res.drawable.description),
+                                    contentDescription = null,
+                                    Modifier.size(16.dp)
+                                )
+                            }
+                        }
+                    }
+
                 }
+
             }
         }
         HorizontalDivider()
@@ -123,14 +186,25 @@ class MCPdirectKeyToolPermissionViewModel : ViewModel() {
     private val _toolMarkerIds = MutableStateFlow(mutableSetOf<Long>())
     val toolMarkers : StateFlow<List<AIPortToolMaker>> = combine(
         ToolRepository.toolMakers,
-        _toolMarkerIds
-        ){ makers,ids ->
-        makers.values.filter { it.id in ids }.toList()
+        _toolMarkerIds,
+                toolMakerFilter
+        ){ makers,ids,filter ->
+        makers.values.filter { (filter.isEmpty()||it.name.lowercase().contains(filter.lowercase()))&&it.id in ids }.toList()
     }.stateIn(
         scope = viewModelScope,
         started = SharingStarted.WhileSubscribed(5000),
         initialValue = emptyList()
     )
+//    val toolMarkers : StateFlow<List<AIPortToolMaker>> = combine(
+//        ToolRepository.toolMakers,
+//        toolMakerFilter
+//    ){ makers,filter ->
+//        makers.values.filter { (filter.isEmpty()||it.name.lowercase().contains(filter.lowercase()))}.toList()
+//    }.stateIn(
+//        scope = viewModelScope,
+//        started = SharingStarted.WhileSubscribed(5000),
+//        initialValue = emptyList()
+//    )
     fun <T : AIPortToolPermission> onLoadToolPermissions(resp: AIPortServiceResponse<List<T>>) {
         if (resp.successful()) resp.data?.let {
             it.forEach {
@@ -139,7 +213,6 @@ class MCPdirectKeyToolPermissionViewModel : ViewModel() {
                 _toolMarkerIds.update { set->
                     set.toMutableSet().apply { add(it.makerId) }
                 }
-//                _toolMarkerIds.value.add(it.makerId)
             }
         }
     }
@@ -148,7 +221,6 @@ class MCPdirectKeyToolPermissionViewModel : ViewModel() {
         if(key!=null&&key.id>Int.MAX_VALUE) {
             _toolPermissions.clear()
             toolPermissions.clear()
-//            _toolMarkerIds.value.clear()
             _toolMarkerIds.update { set->
                 set.toMutableSet().apply { clear() }
             }